avoid duplicate entrypoints when programs depend on programs
