+++ memo/src/lib.rs
@@ -2,7 +2,7 @@
 
 //! A simple program that accepts a string of encoded characters and verifies that it parses. Currently handles UTF-8.
 
-pub mod processor;
+pub mod entrypoint;
 
 // Export current solana-sdk types for downstream users who may also be building with a different
 // solana-sdk version
