+++ token-swap/Cargo.toml
@@ -12,8 +12,9 @@ edition = "2018"
 exclude = ["js/**"]
 
 [features]
+no-entrypoint = []
 skip-no-mangle = ["solana-sdk/skip-no-mangle"]
-program = ["solana-sdk/program", "spl-token/default"]
+program = ["solana-sdk/program", "spl-token/program", "spl-token/no-entrypoint"]
 default = ["solana-sdk/default", "spl-token/default"]
 
 [dependencies]
