+++ token-swap/src/entrypoint.rs
@@ -1,0 +1,30 @@
+//! Program entrypoint definitions
+
+#![cfg(feature = "program")]
+#![cfg(not(feature = "no-entrypoint"))]
+
+use crate::{
+    error::Error,
+    state::State,
+};
+use solana_sdk::{
+    account_info::{AccountInfo},
+    entrypoint,
+    entrypoint::ProgramResult,
+    pubkey::Pubkey,
+    program_error::PrintProgramError
+};
+
+entrypoint!(process_instruction);
+fn process_instruction<'a>(
+    program_id: &Pubkey,
+    accounts: &'a [AccountInfo<'a>],
+    instruction_data: &[u8],
+) -> ProgramResult {
+    if let Err(error) = State::process(program_id, accounts, instruction_data) {
+        // catch the error so we can print it
+        error.print::<Error>();
+        return Err(error);
+    }
+    Ok(())
+}
