+++ token/Cargo.toml
@@ -12,6 +12,7 @@ edition = "2018"
 exclude = ["js/**"]
 
 [features]
+no-entrypoint = []
 skip-no-mangle = ["solana-sdk/skip-no-mangle"]
 program = ["solana-sdk/program"]
 default = ["solana-sdk/default"]
