+++ token/src/entrypoint.rs
@@ -1,0 +1,29 @@
+//! Program entrypoint
+
+#![cfg(feature = "program")]
+#![cfg(not(feature = "no-entrypoint"))]
+
+use crate::{
+    error::TokenError,
+    processor::Processor,
+};
+use solana_sdk::{
+    account_info::{AccountInfo},
+    entrypoint,
+    entrypoint::ProgramResult,
+    pubkey::Pubkey,
+};
+
+entrypoint!(process_instruction);
+fn process_instruction<'a>(
+    program_id: &Pubkey,
+    accounts: &'a [AccountInfo<'a>],
+    instruction_data: &[u8],
+) -> ProgramResult {
+    if let Err(error) = Processor::process(program_id, accounts, instruction_data) {
+        // catch the error so we can print it
+        error.print::<TokenError>();
+        return Err(error);
+    }
+    Ok(())
+}
